requirements.txt
flask
gevent
collections
os
python-dotenv